organize examples in folders
