+++ website/src/examples-sidebar.js
@@ -20,8 +20,8 @@ const sidebars = {
       type: 'category',
       label: '@deck.gl-community/editable-layers',
       items: [
-        "editable-layers-editor",
-        "editable-layers-advanced"
+        "editable-layers/editor",
+        "editable-layers/advanced"
       ]
     },
     // {
