--- website/src/examples/editable-layers-advanced.mdx
@@ -1,5 +1,0 @@
-# Advanced
-
-import Demo from './editable-layers-advanced';
-
-<Demo />
