--- website/src/examples/editable-layers-editor.mdx
@@ -1,5 +1,0 @@
-# Editor
-
-import Demo from './editable-layers-editor';
-
-<Demo />
