+++ website/src/examples/editable-layers/advanced.mdx
@@ -1,0 +1,5 @@
+# Advanced
+
+import Demo from './advanced';
+
+<Demo />
