+++ website/src/examples/editable-layers/editor.mdx
@@ -1,0 +1,5 @@
+# Editor
+
+import Demo from './editor';
+
+<Demo />
